requirements.txt
Flask==2.3.6
flask-cors==3.1.2
pdf2image==1.16.3
pytesseract==0.3.10
openai==1.32.2
python-dotenv==1.0.0
Pillow==10.1.0
